chore: Sprint C mypy strict 清零（150→0）+ CI 硬门槛 + 6 个真 bug 修复
- mesh 加载返回 CVTMesh、裸 dict/list 参数化、Optional 收窄、
  TYPE_CHECKING 导入整理；无 stubs 库以 mypy overrides 抑制（附理由）
- scripts/validate_climate.py 与 import_earth_elevation.py 下沉进包
  （cli_climate 不再 sys.path hack），scripts/ 保留薄壳，调用方式不变
- mypy 抓出的真 bug：worlds.py 分支目录用不存在的 path_name（→name）；
  voronoi_generator legacy 构造 TectonicPlate 缺 euler_pole；三处循环变量
  与文件句柄重名；orographic_precipitation/simulate_climate 返回注解不符；
  importer.py 改用 Image.Resampling.LANCZOS（Pillow 移除旧别名）
- dev extras 补 types-PyYAML；validation extra 去掉与主依赖重复的 tifffile
- tests.yml mypy 翻转硬门槛；roadmap §工程卫生 1/2 项标记已清偿

+++ src/dreamulator/map/noise_kernels.py
@@ -64,7 +64,7 @@ def _lerp(a: float, b: float, t: float) -> float:
 @njit(cache=True)
 def _grad(h: int, x: float, y: float, z: float) -> float:
     g = _GRADS[h & 11]
-    return g[0] * x + g[1] * y + g[2] * z
+    return float(g[0] * x + g[1] * y + g[2] * z)
 
 
 @njit(cache=True)
@@ -105,7 +105,7 @@ def perlin3(x: float, y: float, z: float, perm: np.ndarray) -> float:
 def _noise_points(xs: np.ndarray, ys: np.ndarray, zs: np.ndarray, perm: np.ndarray) -> np.ndarray:
     n = len(xs)
     out = np.empty(n, dtype=np.float64)
-    for i in prange(n):
+    for i in prange(n):  # type: ignore[no-untyped-call,attr-defined]
         out[i] = perlin3(xs[i], ys[i], zs[i], perm)
     return out
 
@@ -123,7 +123,7 @@ def _fbm_points(
 ) -> np.ndarray:
     n = len(xs)
     out = np.empty(n, dtype=np.float64)
-    for i in prange(n):
+    for i in prange(n):  # type: ignore[no-untyped-call,attr-defined]
         total = 0.0
         norm = 0.0
         amp = 1.0
